+++ packages/shared/src/catalog/index.ts
@@ -29,6 +29,7 @@ export * from "./primitives/tabs";
 export * from "./primitives/tagGroup";
 export * from "./primitives/textField";
 export * from "./primitives/timeField";
+export * from "./primitives/tooltip";
 export * from "./primitives/toolbar";
 export * from "./primitives/toggleButton";
 export * from "./primitives/toggleButtonGroup";